requirements.txt
numpy>=1.17.0
torch>=1.2.0
torchvision>=0.4.0
Pillow>=6.1.0
scipy>=1.3.0
statsmodels>=0.10.1
pandas>=0.25.0